배포

+++ src/main/java/site/sac/service/PostServiceImpl.java
@@ -4,11 +4,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.RequestBody;
 import site.sac.dto.PostDTO;
 import site.sac.mapper.PostMapper;
+import site.sac.mapper.UserLikeBoardMapper;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Service
 @Slf4j
@@ -16,33 +18,55 @@ public class PostServiceImpl implements PostService{
     @Autowired
     private PostMapper postMapper;
 
+    @Autowired
+    private UserLikeBoardMapper userLikeBoardMapper;
+
     @Transactional
     @Override
-    public void register(PostDTO postDTO) {
-        log.info("Post : " + postDTO.toString());
+    public void register(PostDTO postDTO, long userId) {
+        postDTO.setUser_id(userId);
         postMapper.insert(postDTO);
     }
 
     @Override
-    public List<PostDTO> getAllPost() {
+    public Map<String,Object> getAllPost() {
         List<PostDTO> allPost = postMapper.getAllPost();
-        allPost.forEach(post->log.info("post : " + post.toString()));
-
-        return allPost;
+        Map<String,Object> result = new HashMap<>();
+        result.put("posts", allPost);
+        return result;
     }
 
     @Override
     public PostDTO getPostDetail(Long postId) {
-        PostDTO post = postMapper.read(postId);
-        return post;
+        return postMapper.read(postId);
     }
 
     @Override
-    public List<PostDTO> getPostsByBoardId(Long boardId) {
+    public Map<String,Object> getPostsByBoardId(Long boardId) {
         List<PostDTO> postsByBoardId = postMapper.getAllPostsByBoardId(boardId);
-        postsByBoardId.forEach(post->log.info("post : " + post.toString()));
 
-        return postsByBoardId;
+        if (postsByBoardId == null){
+            throw new NullPointerException();
+        }
+
+        Map<String,Object> result = new HashMap<>();
+        result.put("posts", postsByBoardId);
+        result.put("count", postsByBoardId.size());
+
+        return result;
+    }
+    @Override
+    public Map<String,Object> getAllPostByUserId(Long userId){
+        List<String> userLikeBoard = userLikeBoardMapper.getAllByUserId(userId);
+        if (userLikeBoard == null){
+            throw new NullPointerException();
+        }
+        List<PostDTO> posts = postMapper.getAllPostByUserLikeBoard(userLikeBoard);
+        Map<String,Object> result = new HashMap<>();
+        result.put("posts", posts);
+        result.put("count", posts.size());
+        return result;
+
     }
 
     @Override
@@ -53,20 +77,25 @@ public class PostServiceImpl implements PostService{
     }
 
     @Override
-    public PostDTO postEdit(PostDTO postDTO) {
-        postMapper.update(postDTO);
-        return postDTO;
+    public void postEdit(PostDTO postDTO, long postId) {
+        if(postMapper.read(postId).getUser_id() == postDTO.getUser_id()){
+            postMapper.update(postDTO);
+        } else {
+            throw new NullPointerException();
+        }
     }
 
     @Override
-    public void delete(Long postId) {
-        postMapper.delete(postId);
+    public void delete(PostDTO postDTO, long postId) {
+        if(postMapper.read(postId).getUser_id()== postDTO.getUser_id()){
+            postMapper.delete(postId);
+        } else {
+            throw new NullPointerException();
+        }
     }
 
     @Override
     public List<PostDTO> getPostsByLike(List<Long> likes) {
-
         return postMapper.getPostsByLike(likes);
     }
-
 }
